Generator model
Ok, so we got a generator model no. So far LSTm version because thats a little easier to handle for my puny brain.

ToDos:
Add L2 Reg for LSTM
Add a tanh activation for the Attention, so the inputs all have a similar scalre
Add proper dropout to the attention tingie
eh, dunno