Now with better results
changed dropout in the lstm to use recurrent dropout and ndormal dropout ... cpu version is slower but this actually regularizes... ahaha
Changed the dropout for the rest of the network to be random for each step.

added a generator encoder and a fixed generator, both of which follow a blockwise residual structure now.