Experiment Setup
For use with tensorboard copy all the runs into a single folder, I have not automated this part yet.
This also makes a neato home for the metrics pickle that is created anyways, so its better organized n shit.
Tensorboard can then be called on this folder and will give you neato graphz

Otherwise, analysis is pretty straightforward